// web-application/src/types-openapi/api/models/UpdateServiceRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateServiceRequest
 */
export interface UpdateServiceRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceRequest
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceRequest
     */
    code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceRequest
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceRequest
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceRequest
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateServiceRequest
     */
    linkedProductId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateServiceRequest
     */
    pricePerHour?: number | null;
}

/**
 * Check if a given object implements the UpdateServiceRequest interface.
 */
export function instanceOfUpdateServiceRequest(value: object): value is UpdateServiceRequest {
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function UpdateServiceRequestFromJSON(json: any): UpdateServiceRequest {
    return UpdateServiceRequestFromJSONTyped(json, false);
}

export function UpdateServiceRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateServiceRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'code': json['code'] == null ? undefined : json['code'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'linkedProductId': json['linkedProductId'] == null ? undefined : json['linkedProductId'],
        'pricePerHour': json['pricePerHour'] == null ? undefined : json['pricePerHour'],
    };
}

export function UpdateServiceRequestToJSON(json: any): UpdateServiceRequest {
    return UpdateServiceRequestToJSONTyped(json, false);
}

export function UpdateServiceRequestToJSONTyped(value?: UpdateServiceRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'code': value['code'],
        'name': value['name'],
        'description': value['description'],
        'imageUrl': value['imageUrl'],
        'linkedProductId': value['linkedProductId'],
        'pricePerHour': value['pricePerHour'],
    };
}